refactor: eatinordertables 패키지로 OrderTable 관련 클래스 분리
OrderTable과 관련된 application, domain, ui 를 eatinordertables 로 옮김
