+++ src/main/java/kitchenpos/eatinordertables/domain/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinordertables.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
